Cleansing/requirements.txt
mysql-connector-python==8.0.21
pandas==1.1.2
SQLAlchemy==1.3.19
Unidecode==1.1.1
geopy==2.0.0
PyMySQL==0.10.1
mysql.connector==2.2.9
cryptography==3.1.1
prettytable==2.0.0
tqdm==4.56.0
tabulate==0.8.9
findspark==1.4.2
pyspark==3.0.1
python-dateutil==2.7.3